+++ hazelcast/src/test/java/com/hazelcast/map/impl/query/MapDisableCopyOnReadTest.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2008-2020, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.hazelcast.map.impl.query;
 
 import com.hazelcast.config.Config;
@@ -5,14 +21,11 @@ import com.hazelcast.config.InMemoryFormat;
 import com.hazelcast.config.MapConfig;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.PartitionService;
-import com.hazelcast.map.Immutable;
 import com.hazelcast.map.impl.proxy.MapProxyImpl;
 import com.hazelcast.test.HazelcastTestSupport;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-
-import java.io.Serializable;
 
 public class MapDisableCopyOnReadTest extends HazelcastTestSupport {
 
@@ -26,8 +39,8 @@ public class MapDisableCopyOnReadTest extends HazelcastTestSupport {
         Config config = new Config().addMapConfig(new MapConfig("myMap*").setInMemoryFormat(InMemoryFormat.OBJECT));
         instance = createHazelcastInstance(config);
         partitionService = instance.getPartitionService();
-        immutableMapProxy = (MapProxyImpl)instance.getMap("myMapImmutable");
-        mutableMapProxy = (MapProxyImpl)instance.getMap("myMapMutable");
+        immutableMapProxy = (MapProxyImpl) instance.getMap("myMapImmutable");
+        mutableMapProxy = (MapProxyImpl) instance.getMap("myMapMutable");
     }
 
     @Test
@@ -72,30 +85,15 @@ public class MapDisableCopyOnReadTest extends HazelcastTestSupport {
         Assert.assertFalse(result1 == result2);
     }
 
-}
+    @Test
+    public void testClassCastException() {
+        String key = "testCorKey";
+        TestEntityImmutable tc = new TestEntityImmutable("hello");
+        immutableMapProxy.put(key, tc);
+        int partitionId = partitionService.getPartition(key).getPartitionId();
 
-class TestEntityImmutable implements Immutable, Serializable {
-    private static final long serialVersionUID = 1L;
-    private String test;
-
-    public TestEntityImmutable(String test) {
-        this.test = test;
-    }
-
-    public String getTest() {
-        return test;
-    }
-}
-
-class TestEntity implements Serializable {
-    private static final long serialVersionUID = 1L;
-    private String test;
-
-    public TestEntity(String test) {
-        this.test = test;
-    }
-
-    public String getTest() {
-        return test;
+        TestEntityImmutable result1 =  immutableMapProxy.iterator(10, partitionId, false).next().getValue();
+        TestEntityImmutable result2 =  immutableMapProxy.iterator(10, partitionId, false).next().getValue();
+        Assert.assertTrue(result1 == result2);
     }
 }
