+++ hazelcast/src/test/java/com/hazelcast/map/impl/query/TestEntity.java
@@ -14,10 +14,19 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map;
+package com.hazelcast.map.impl.query;
 
-/**
- * This is just a marker interface for Adobe Entities so that we can disable Copy on Read for this type
- */
-public interface Immutable {
+import java.io.Serializable;
+
+class TestEntity implements Serializable {
+    private static final long serialVersionUID = 1L;
+    private String test;
+
+    public TestEntity(String test) {
+        this.test = test;
+    }
+
+    public String getTest() {
+        return test;
+    }
 }
